restructure sync code
